Initialize README
Review notes and assignments above.
I will have the data pre-processing and write-up done by Friday.  Much of the individual write-ups for tests can be down without the data (such as why we're performing tests, test theory, and assumptions), but you can also just import the 2000's csv and perform your analysis of the raw data.  If the dataset changes due to pre-processing, you can just re-run (but not re-code) the tests with the new data and update conclusion if those change.